+++ packages/elements/icon/src/index.js
@@ -1,0 +1,5 @@
+// @flow
+import Icon, { size } from './components/Icon';
+
+export default Icon;
+export { size };