+++ packages/elements/inline-message/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+// eslint-disable-next-line import/export
+export * from './src';