+++ packages/elements/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/breadcrumbs
 
+## 3.1.10
+
+- [patch] Update types [2fe5453](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/2fe5453)
+
 ## 3.1.9
 
 ## 3.1.8