+++ packages/fabric/media-card/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/media-image
 
+## 23.1.1
+
+- [patch] Update dependencies [623f8ca](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/623f8ca)
+
 ## 23.1.0
 
 - [minor] Added new AppCardView v1.5 designs behind a feature flag. [92bc6c8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/92bc6c8)